requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
requests
httpx
pydantic
